requirements.txt
flask
flask_cors
waitress
requests
Pillow
PyMuPDF
easyocr